# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Digital Wallet API",
        "description": "API RESTful para carteira digital pessoal. Permite autenticação, depósitos, saques e consulta de histórico financeiro.",
        "contact": {
            "email": "dev@digitalwallet.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor principal"
        }
    ],
    "paths": {
        "/api/accounts": {
            "get": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Listar contas do usuário",
                "operationId": "13efa5706d1fae14fc1cf87c9f6b1427",
                "responses": {
                    "200": {
                        "description": "Contas listadas com sucesso"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Criar nova conta",
                "operationId": "e974666bc6c050296a8ee09a3ab6533f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "type"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Nubank"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "checking",
                                            "savings",
                                            "cash",
                                            "investment"
                                        ]
                                    },
                                    "balance": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 0
                                    },
                                    "color": {
                                        "type": "string",
                                        "example": "#8A05BE"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Conta criada com sucesso"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/accounts/{account}": {
            "delete": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Arquivar conta",
                "description": "A conta é arquivada (soft), não excluída - preserva o histórico de transações.",
                "operationId": "a30295a60453fd97c43c5ffd510e9858",
                "parameters": [
                    {
                        "name": "account",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Conta arquivada com sucesso"
                    },
                    "404": {
                        "description": "Conta não encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Atualizar conta",
                "operationId": "060f4c0a36626cfe412a32a49a70ed27",
                "parameters": [
                    {
                        "name": "account",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Conta atualizada com sucesso"
                    },
                    "404": {
                        "description": "Conta não encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}